[adapters] rename to decorator
