+++ test/CMakeLists.txt
@@ -181,5 +181,5 @@ add_subdirectory(grid)
 add_subdirectory(vtk)
 add_subdirectory(xml)
 add_subdirectory(api)
-add_subdirectory(adapters)
+add_subdirectory(decorators)
 add_subdirectory(traits)
